* Automatically adds http.keepalive.timeout when ws.inactivity.timeout is specified for connect options * Cleaned up DefaultConnectOptionsContext
