+++ server/src/test/java/org/kaazing/gateway/server/config/parse/ServiceDefaultsTest.java
@@ -71,11 +71,11 @@ public class ServiceDefaultsTest {
         Assert.assertTrue("en0".equals(connectOptionsMap.get("udp.interface")));
         System.out.println(connectOptionsMap.get("tcp.transport"));
         Assert.assertTrue("socks://localhost:8000".equals(connectOptionsMap.get("tcp.transport").toString().trim()));
-        Assert.assertTrue("disabled".equalsIgnoreCase((String) connectOptionsMap.get("ssl.encryption")));
+        Assert.assertFalse((Boolean)connectOptionsMap.get("ssl.encryptionEnabled"));
 
-        Assert.assertEquals(5, connectOptionsMap.get(HTTP_KEEP_ALIVE_TIMEOUT_KEY));
-        Assert.assertFalse((Boolean) connectOptionsMap.get(HTTP_KEEP_ALIVE));
-        Assert.assertEquals(7, connectOptionsMap.get("http.keepalive.connections"));
+        Assert.assertEquals(5, connectOptionsMap.get("http[http/1.1]."+HTTP_KEEP_ALIVE_TIMEOUT_KEY));
+        Assert.assertFalse((Boolean) connectOptionsMap.get("http[http/1.1]."+HTTP_KEEP_ALIVE));
+        Assert.assertEquals(7, connectOptionsMap.get("http[http/1.1].keepalive.connections"));
     }
 
     @Test
